Fix excessive rows when selecting from active session history
