+++ pkg/collector/corechecks/oracle/oracle.go
@@ -113,7 +113,7 @@ type Check struct {
 	openMode                                string
 	legacyIntegrationCompatibilityMode      bool
 	clock                                   clock.Clock
-	lastSampleId                            uint64
+	lastSampleID                            uint64
 }
 
 type vDatabase struct {
